refactor: improve timer logic and UI components
This commit introduces several refinements:

- Renamed `questionSecondsLimit` to `questionTimeLimitSeconds` in `BaseGameInteractionListener.kt` and `BaseGameViewModel.kt` for clarity.
- Adjusted the `TimerComponent.kt`:
    - Changed `Canvas` size from `48.dp` to `40.dp`.
    - Changed `strokeWidth` from `5.dp` to `3.dp`.
- Moved `GuessCard.kt` from `com.amsterdam.ui.components.guessGame` to `com.amsterdam.ui.screens.letsPlay.component`.
- Updated imports in `GuessPicture.kt` and `GuessTitle.kt` to reflect the new location of `GuessCard`.
