+++ viewModel/src/main/java/com/amsterdam/viewmodel/sharedGame/BaseGameInteractionListiner.kt
@@ -2,7 +2,7 @@ package com.amsterdam.viewmodel.sharedGame
 
 interface BaseGameInteractionListener {
     fun startTheTimer(
-        questionSecondsLimit: Int,
+        questionTimeLimitSeconds: Int,
         onTimerChange: (BaseGameUiState) -> Unit,
         onTimerFinished: () -> Unit
     )
